+++ lib/receipt_verifier.ex
@@ -3,42 +3,45 @@ defmodule ReceiptVerifier do
   Verify iTunes receipt with the Apple Store
 
   ## Example
-      iex> {:ok, receipt} = ReceiptVerifier.verify(base64_encoded_receipt_data)
-      ...> receipt =
-        %ReceiptVerifier.Receipt{receipt: {"adam_id" => 0, "app_item_id" => 0, "application_version" => "1241",
-          "bundle_id" => "com.sumiapp.GridDiary", "download_id" => 0,
-          "in_app" => [%{"is_trial_period" => "false",
-             "original_purchase_date" => "2014-08-04 06:24:51 Etc/GMT",
-             "original_purchase_date_ms" => "1407133491000",
-             "original_purchase_date_pst" => "2014-08-03 23:24:51 America/Los_Angeles",
-             "original_transaction_id" => "1000000118990828",
-             "product_id" => "com.sumiapp.GridDiary.pro",
-             "purchase_date" => "2014-09-02 03:29:06 Etc/GMT",
-             "purchase_date_ms" => "1409628546000",
-             "purchase_date_pst" => "2014-09-01 20:29:06 America/Los_Angeles",
-             "quantity" => "1", "transaction_id" => "1000000118990828"},
-           %{"is_trial_period" => "false",
-             "original_purchase_date" => "2014-09-02 03:29:06 Etc/GMT",
-             "original_purchase_date_ms" => "1409628546000",
-             "original_purchase_date_pst" => "2014-09-01 20:29:06 America/Los_Angeles",
-             "original_transaction_id" => "1000000122102348",
-             "product_id" => "com.sumiapp.griddiary.test",
-             "purchase_date" => "2014-09-02 03:29:06 Etc/GMT",
-             "purchase_date_ms" => "1409628546000",
-             "purchase_date_pst" => "2014-09-01 20:29:06 America/Los_Angeles",
-             "quantity" => "1", "transaction_id" => "1000000122102348"}],
-          "original_application_version" => "1.0",
-          "original_purchase_date" => "2013-08-01 07:00:00 Etc/GMT",
-          "original_purchase_date_ms" => "1375340400000",
-          "original_purchase_date_pst" => "2013-08-01 00:00:00 America/Los_Angeles",
-          "receipt_creation_date" => "2014-09-02 03:29:06 Etc/GMT",
-          "receipt_creation_date_ms" => "1409628546000",
-          "receipt_creation_date_pst" => "2014-09-01 20:29:06 America/Los_Angeles",
-          "receipt_type" => "ProductionSandbox",
-          "request_date" => "2016-04-29 07:52:28 Etc/GMT",
-          "request_date_ms" => "1461916348197",
-          "request_date_pst" => "2016-04-29 00:52:28 America/Los_Angeles",
-          "version_external_identifier" => 0}}
+      iex> ReceiptVerifier.verify(base64_encoded_receipt_data)
+      ...> {:ok, %ReceiptVerifier.ResponseData{app_receipt: %ReceiptVerifier.AppReceipt{adam_id: 0,
+        app_item_id: 0, application_version: "1241",
+        bundle_id: "com.sumiapp.GridDiary", download_id: 0,
+        in_app: [%ReceiptVerifier.IAPReceipt{expires_date: nil,
+          is_trial_period: false,
+          original_purchase_date: %DateTime{calendar: Calendar.ISO, day: 17, hour: 6,
+           microsecond: {491000, 6}, minute: 52, month: 1, second: 13, std_offset: 0,
+           time_zone: "Etc/UTC", utc_offset: 0, year: 1970, zone_abbr: "UTC"},
+          original_transaction_id: "1000000118990828",
+          product_id: "com.sumiapp.GridDiary.pro",
+          purchase_date: %DateTime{calendar: Calendar.ISO, day: 17, hour: 7,
+           microsecond: {546000, 6}, minute: 33, month: 1, second: 48, std_offset: 0,
+           time_zone: "Etc/UTC", utc_offset: 0, year: 1970, zone_abbr: "UTC"},
+          quantity: 1, transaction_id: "1000000118990828",
+          web_order_line_item_id: nil},
+         %ReceiptVerifier.IAPReceipt{expires_date: nil, is_trial_period: false,
+          original_purchase_date: %DateTime{calendar: Calendar.ISO, day: 17, hour: 7,
+           microsecond: {546000, 6}, minute: 33, month: 1, second: 48, std_offset: 0,
+           time_zone: "Etc/UTC", utc_offset: 0, year: 1970, zone_abbr: "UTC"},
+          original_transaction_id: "1000000122102348",
+          product_id: "com.sumiapp.griddiary.test",
+          purchase_date: %DateTime{calendar: Calendar.ISO, day: 17, hour: 7,
+           microsecond: {546000, 6}, minute: 33, month: 1, second: 48, std_offset: 0,
+           time_zone: "Etc/UTC", utc_offset: 0, year: 1970, zone_abbr: "UTC"},
+          quantity: 1, transaction_id: "1000000122102348",
+          web_order_line_item_id: nil}], original_application_version: "1.0",
+        original_purchase_date: %DateTime{calendar: Calendar.ISO, day: 16, hour: 22,
+         microsecond: {400000, 6}, minute: 2, month: 1, second: 20, std_offset: 0,
+         time_zone: "Etc/UTC", utc_offset: 0, year: 1970, zone_abbr: "UTC"},
+        receipt_creation_date: %DateTime{calendar: Calendar.ISO, day: 17, hour: 7,
+         microsecond: {546000, 6}, minute: 33, month: 1, second: 48, std_offset: 0,
+         time_zone: "Etc/UTC", utc_offset: 0, year: 1970, zone_abbr: "UTC"},
+        receipt_type: "ProductionSandbox",
+        request_date: %DateTime{calendar: Calendar.ISO, day: 18, hour: 2,
+         microsecond: {590831, 6}, minute: 47, month: 1, second: 30, std_offset: 0,
+         time_zone: "Etc/UTC", utc_offset: 0, year: 1970, zone_abbr: "UTC"},
+        version_external_identifier: 0}, base64_latest_app_receipt: nil,
+       latest_iap_receipts: []}}
 
   > Note: If you send sandbox receipt to production server, it will be auto resend to test server. Same for the production receipt.
   """
@@ -53,7 +56,7 @@ defmodule ReceiptVerifier do
   def verify(receipt) when is_binary(receipt) do
     with(
       {:ok, json} <- Client.request(receipt),
-      {:ok, data} <- Parser.parse(json)
+      {:ok, data} <- Parser.parse_response(json)
     ) do
       {:ok, data}
     else